Move all settings to better layout
