+++ app/Http/Controllers/SettingsController.php
@@ -13,15 +13,6 @@ use Illuminate\Validation\Rule;
  */
 class SettingsController extends Controller
 {
-    /**
-     * @deprecated
-     */
-    public function renderFollowerSettings(): Renderable {
-        return view('settings.follower', [
-            'requests'  => auth()->user()->followRequests()->with('user')->paginate(15),
-            'followers' => auth()->user()->followers()->with('user')->paginate(15)
-        ]);
-    }
 
     /**
      * @deprecated
