+++ resources/views/settings/account.blade.php
@@ -1,0 +1,7 @@
+@extends('layouts.settings')
+
+
+@section('content')
+    @include('settings.cards.password')
+    @include('settings.cards.account-deletion')
+@endsection
